+++ packages/core/src/services/stubs/policy.service.stub.js
@@ -1,8 +1,0 @@
-export const policyService = {
-    getPolicy: async (_policyId) => {
-        return null;
-    },
-    evaluateAccess: (_policy, _toolName) => {
-        return true;
-    }
-};