+++ packages/core/src/services/stubs/toon.serializer.stub.js
@@ -1,6 +1,0 @@
-"use strict";
-Object.defineProperty(exports, "__esModule", { value: true });
-exports.toonSerializer = void 0;
-exports.toonSerializer = {
-    serialize: function (_data) { return JSON.stringify(_data); },
-};